+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 21.09.0 (in progress)
+
+FEATURES
+* Producer API: C client 
+
+IMPROVEMENTS
+* Allow using ASAPO for commissioning beamtimes
+
 ## 21.06.0
 
 FEATURES